feat: 내 모임 조회 및 모임box 컴포넌트화 #150
