+++ src/app/(meetUp)/meetup/page.tsx
@@ -3,6 +3,7 @@ import { cookies } from 'next/headers'
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs'
 import styles from './meetup.module.scss'
 import MeetupTabPage from './_components/MeetupTabPage'
+import { getMeetupList } from '@/utils/apiRequest/meetupApiRequest'
 
 const MeetupPage = async () => {
   const cookieStore = cookies()
@@ -21,6 +22,7 @@ const MeetupPage = async () => {
    * 생성된 모임에서 보여줄 게시물은 meetup-board 데이터 받아서 조회 하도록
    * 생성된 모임에 영상 보러가기 버튼 클릭 하면 해당 게시물의 디테일 페이지로 이동
    */
+  const meetupList = await getMeetupList()
 
   return (
     <div style={{ paddingTop: '4.6619rem' }}>
@@ -33,7 +35,7 @@ const MeetupPage = async () => {
 
       <div className={styles.container}>
         <div className={`inner-box ${styles.contents}`}>
-          <MeetupTabPage />
+          <MeetupTabPage meetupList={meetupList} />
         </div>
       </div>
     </div>
